refactor(shared): services files dialog search all (#168)
* refactor(shared): service get search

* refactor(shared): service get search

* refactor(shared): interfaces

* refactor(shared): name folder service

* refactor(shared): absolute import

* refactor(shared): file schema

* refactor(shared): datagrid column definition

By following the documentation: The columns prop should keep the same reference between two renders. The columns are designed to be definitions, to never change once the component is mounted. Otherwise, you take the risk of losing elements like column width or order. You can create the array outside the render function or memoize it.

* refactor(shared): absolute import